move /user and /users routes from app to routers
