+++ routers/tags.js
@@ -1,5 +1,4 @@
 const { Router } = require("express")
-const mongoose = require("mongoose")
 const { Tag } = require("../models/tag")
 
 const router = Router()
